Collections of data
python provides a range of tools and libraries for building hand gesture recognition systems, from simple systems based on OpenCV to more complex systems based on machine learning and deep learning. With the right tools and techniques, it is possible to build powerful gesture recognition systems that can be used in a range of applications, from gaming and virtual reality to sign language recognition and human-computer interaction.